docker-processing/src/processing_protos/requirements.txt
configobj==5.0.6
grpcio==1.36.1
grpcio-reflection==1.36.1
grpcio-tools==1.36.1
protobuf==3.15.5
six==1.15.0